requirements.txt
python-telegram-bot[http2]
supabase
python-dotenv